refactor(cli): create cli/ sub-package — 5 files (slice 7/8, #170)
