+++ src/voicecli/cli/__init__.py
@@ -1,0 +1,3 @@
+from voicecli.cli.main import app
+
+__all__ = ["app"]
