+++ src/voicecli/cli/doctor.py
