+++ tools/file_exemptions.txt
@@ -1,8 +1,8 @@
 # Exemptions — file_length gate
 # Format: <path> <issue-url>  (single-space separator)
 # Each entry MUST reference a tracking issue so the exemption is recoverable.
-src/voicecli/cli.py https://github.com/Roxabi/voiceCLI/issues/129
-src/voicecli/cli_dictate.py https://github.com/Roxabi/voiceCLI/issues/129
+src/voicecli/cli/main.py https://github.com/Roxabi/voiceCLI/issues/129
+src/voicecli/cli/dictate.py https://github.com/Roxabi/voiceCLI/issues/129
 src/voicecli/api/api.py https://github.com/Roxabi/voiceCLI/issues/81
 src/voicecli/runtime/stt_daemon.py https://github.com/Roxabi/voiceCLI/issues/82
 src/voicecli/runtime/wire_protocol.py https://github.com/Roxabi/voiceCLI/issues/83
